user: move event to events mod
