+++ src/events/mod.rs
@@ -28,3 +28,4 @@ pub(crate) mod ovs;
 pub(crate) mod skb;
 pub(crate) mod skb_drop;
 pub(crate) mod skb_tracking;
+pub(crate) mod user;
